Move header processing into blockheaderprocessor.h
