+++ src/utilprocessmsg.h
@@ -8,15 +8,4 @@ class CBlockHeader;
 
 bool HaveBlockData(const uint256& hash);
 
-// Process received block header.
-class BlockHeaderProcessor {
-public:
-    // returns false on error
-    virtual bool operator()(const std::vector<CBlockHeader>& headers,
-            bool peerSentMax, bool maybeAnnouncement) = 0;
-
-    virtual ~BlockHeaderProcessor() = 0;
-};
-inline BlockHeaderProcessor::~BlockHeaderProcessor() { }
-
 #endif
